Fix-database operation(Cascade delete)

+++ src/main/java/pl/jbed/stud/SomeWebService/Entity/Role.java
@@ -9,13 +9,12 @@ public class Role {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name="id")
-    private long id;
+    private int id;
 
     @Column(name="name")
     private String name;
 
     public Role(){
-
     }
 
     public Role(String name) {
@@ -26,7 +25,7 @@ public class Role {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
